Refactor: Member 도메인 경로 수정
