+++ src/main/java/com/calefit/crew/CrewService.java
@@ -3,8 +3,8 @@ package com.calefit.crew;
 import com.calefit.common.dto.CommonDtoList;
 import com.calefit.crew.dto.*;
 import com.calefit.crew.entity.Crew;
-import com.calefit.member.member.MemberRepository;
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.MemberRepository;
+import com.calefit.member.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
