+++ src/main/java/com/calefit/inbody/entity/Inbody.java
@@ -3,7 +3,7 @@ package com.calefit.inbody.entity;
 import static javax.persistence.FetchType.*;
 
 import com.calefit.inbody.domain.BodyComposition;
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.entity.Member;
 import java.time.LocalDateTime;
 import javax.persistence.Embedded;
 import javax.persistence.Entity;
