+++ src/main/java/com/calefit/member/MemberRepository.java
@@ -1,6 +1,6 @@
-package com.calefit.member.member;
+package com.calefit.member;
 
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
