+++ src/main/java/com/calefit/member/domain/CrewInfo.java
@@ -1,4 +1,4 @@
-package com.calefit.member.member.domain;
+package com.calefit.member.domain;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
