+++ src/main/java/com/calefit/member/dto/MemberSearchResponse.java
@@ -1,6 +1,6 @@
-package com.calefit.member.member.dto;
+package com.calefit.member.dto;
 
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.entity.Member;
 import lombok.Getter;
 
 @Getter
