+++ src/main/java/com/calefit/member/exception/NotAvailableMemberEmailException.java
@@ -1,9 +1,9 @@
-package com.calefit.member.member.exception;
+package com.calefit.member.exception;
 
 import com.calefit.exception.BusinessException;
 import org.springframework.http.HttpStatus;
 
-public class NotAvailableMemberNicknameException extends BusinessException {
+public class NotAvailableMemberEmailException extends BusinessException {
 
     @Override
     public HttpStatus getHttpStatus() {
