+++ src/main/java/com/calefit/member/exception/NotAvailableMemberLoginException.java
@@ -1,10 +1,9 @@
-package com.calefit.member.member.exception;
+package com.calefit.member.exception;
 
 import com.calefit.exception.BusinessException;
 import org.springframework.http.HttpStatus;
 
-public class NotAvailableMemberEmailException extends BusinessException {
-
+public class NotAvailableMemberLoginException extends BusinessException {
     @Override
     public HttpStatus getHttpStatus() {
         return HttpStatus.BAD_REQUEST;
