+++ src/main/java/com/calefit/member/exception/NotAvailableMemberNicknameException.java
@@ -1,9 +1,10 @@
-package com.calefit.member.member.exception;
+package com.calefit.member.exception;
 
 import com.calefit.exception.BusinessException;
 import org.springframework.http.HttpStatus;
 
-public class NotAvailableMemberLoginException extends BusinessException {
+public class NotAvailableMemberNicknameException extends BusinessException {
+
     @Override
     public HttpStatus getHttpStatus() {
         return HttpStatus.BAD_REQUEST;
