+++ src/main/java/com/calefit/member/exception/NotFoundAlgorithmException.java
@@ -1,12 +1,11 @@
-package com.calefit.member.member.exception;
+package com.calefit.member.exception;
 
 import com.calefit.exception.BusinessException;
 import org.springframework.http.HttpStatus;
 
-public class NotFoundMemberException extends BusinessException {
-
+public class NotFoundAlgorithmException extends BusinessException {
     @Override
     public HttpStatus getHttpStatus() {
-        return HttpStatus.NOT_FOUND;
+        return HttpStatus.INTERNAL_SERVER_ERROR;
     }
 }
