--- src/main/java/com/calefit/member/member/exception/NotFoundAlgorithmException.java
@@ -1,11 +1,0 @@
-package com.calefit.member.member.exception;
-
-import com.calefit.exception.BusinessException;
-import org.springframework.http.HttpStatus;
-
-public class NotFoundAlgorithmException extends BusinessException {
-    @Override
-    public HttpStatus getHttpStatus() {
-        return HttpStatus.INTERNAL_SERVER_ERROR;
-    }
-}
