+++ src/main/java/com/calefit/template/entity/Template.java
@@ -2,7 +2,7 @@ package com.calefit.template.entity;
 
 import static javax.persistence.FetchType.*;
 
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.entity.Member;
 import java.util.ArrayList;
 import java.util.List;
 import javax.persistence.Entity;
