+++ src/main/java/com/calefit/workout/entity/Schedule.java
@@ -2,7 +2,7 @@ package com.calefit.workout.entity;
 
 import static javax.persistence.FetchType.*;
 
-import com.calefit.member.member.entity.Member;
+import com.calefit.member.entity.Member;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
